+++ src/steps/resource-manager/interservice/index.ts
@@ -74,6 +74,7 @@ export async function buildComputeNetworkRelationships(
           continue;
         }
 
+        let subnetId: string | undefined;
         for (const ipConfiguration of nic.ipConfigurations || []) {
           if (ipConfiguration.publicIPAddress) {
             const publicIpEntity = await jobState.findEntity(
@@ -101,30 +102,32 @@ export async function buildComputeNetworkRelationships(
             }
           }
 
-          if (ipConfiguration.subnet) {
-            const subnetEntity = await jobState.findEntity(
-              ipConfiguration.subnet.id as string,
-            );
-            if (subnetEntity) {
-              await jobState.addRelationship(
-                createDirectRelationship({
-                  from: subnetEntity,
-                  _class: InterserviceRelationships.SUBNET_HAS_VM._class,
-                  to: vmEntity,
-                  properties: {
-                    _type: InterserviceRelationships.SUBNET_HAS_VM._type,
-                  },
-                }),
-              );
-            } else {
-              logger.warn(
-                {
-                  vmId: vmData.id,
-                  subnetId: ipConfiguration.subnet.id,
+          if (ipConfiguration.subnet && !subnetId) {
+            subnetId = ipConfiguration.subnet.id;
+          }
+        }
+
+        if (subnetId) {
+          const subnetEntity = await jobState.findEntity(subnetId);
+          if (subnetEntity) {
+            await jobState.addRelationship(
+              createDirectRelationship({
+                from: subnetEntity,
+                _class: InterserviceRelationships.SUBNET_HAS_VM._class,
+                to: vmEntity,
+                properties: {
+                  _type: InterserviceRelationships.SUBNET_HAS_VM._type,
                 },
-                'Could not find subnet in job state',
-              );
-            }
+              }),
+            );
+          } else {
+            logger.warn(
+              {
+                vmId: vmData.id,
+                subnetId: subnetId,
+              },
+              'Could not find subnet in job state',
+            );
           }
         }
       }